Added Ships! -Added three different ships -Reworked GunCycler to be better -Added debug mode (press G) -Switched from requestAnimationFrame to setInterval to improve FPS

+++ script/hostile.js
@@ -20,7 +20,7 @@ var Hostile = Entity.extend(function(props){
 				this.destroy();
 			}
 
-			if (this.collidesCircles(Game.player) && !this.dead) {
+			if (this.collidesCircles(Game.player) && !this.dead && !Game.debugMode) {
 				this.kill();
 				Game.player.kill();
 			}
